refactor(runbotics-orchestrator-ui): :recycle: Extract code do different components; removes max-line-per-function linter issue
